+++ QEfficient/src/__init__.py
@@ -5,4 +5,5 @@
 #
 # -----------------------------------------------------------------------------
 
-from QEfficient.loader.loader import QEFFAutoModel  # noqa: F401
+from QEfficient.src._transformers.auto import QEffAutoModel, QEFFAutoModelForCausalLM  # noqa: F401
+from QEfficient.src.common import QEFFCommonLoader  # noqa: F401
