+++ tests/test_loader.py
@@ -11,8 +11,7 @@ import pytest
 from transformers.models.gpt2.modeling_gpt2 import GPT2LMHeadModel
 
 import QEfficient
-from QEfficient import QEFFAutoModel
-from QEfficient.loader.loader_factory import QEFFAutoModelForCausalLM
+from QEfficient import QEFFAutoModelForCausalLM, QEFFCommonLoader
 
 model_name_to_params_dict : Dict[str, Dict[str, Any]] = {
     "gpt2": {
@@ -27,7 +26,7 @@ model_names = model_name_to_params_dict.keys()
 #FIXME: Add test cases for passing cache_dir, pretrained_model_path instead of card name, etc., Passing other kwargs
 @pytest.mark.parametrize("model_name", model_names)
 def test_qeff_auto_model_for_causal_lm(model_name: str):
-    model = QEFFAutoModel.from_pretrained(model_name)
+    model = QEFFCommonLoader.from_pretrained(model_name)
     assert isinstance(model, model_name_to_params_dict[model_name]['qeff_class'])
     assert isinstance(model.model, model_name_to_params_dict[model_name]['hf_class']) # type: ignore
